jwt,authiritazion

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 
 
@@ -19,18 +21,23 @@ blogsRouter.get('/:id',async(res,req) =>{
     }
 })
 
+
+
 blogsRouter.post('/', async(request, response) => {
     const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
 
     if(body.title === undefined || body.url === undefined){
         return response.status(400).json({error:'title and url are required'})
     }
 
-    const user = await User.findById(body.userId)
+    const user = await User.findById(decodedToken.id)
     console.log(user)
-    if(!user){
-        return response.status(400).json({error:'user does not exist'})
-    }
 
     const blog = new Blog({
         title: body.title,
@@ -47,8 +54,19 @@ blogsRouter.post('/', async(request, response) => {
 } )
 
 blogsRouter.delete('/:id', async(request, response, next) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() === decodedToken.id.toString()){
+
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }else{
+        response.status(401).json({ error: 'token missing or invalid' })
+    }
 } )
 
 blogsRouter.put('/:id', async(request, response) => {
